+++ KlonVampireSurvivors/Assets/Scripts/Weapon/WeaponScriptable.cs
@@ -22,8 +22,8 @@ public class WeaponScriptable : ScriptableObject
     public float duration { get => Duration; private set => Duration = value; }
     
     [SerializeField]
-    int Damage;
-    public int damage { get => Damage; private set => Damage = value; }
+    float Damage;
+    public float damage { get => Damage; private set => Damage = value; }
 
     [SerializeField]
     int Pierce;
